+++ src/libdecaf/src/modules/snd_core/snd_core_voice.cpp
@@ -181,7 +181,7 @@ AXGetVoiceOffsets(AXVoice *voice,
       offsets->endOffset = pageAddress + extras->data.endOffsetAbs - baseAddress;
       offsets->currentOffset = pageAddress + extras->data.currentOffsetAbs - baseAddress;
    } else {
-      be_val<uint8_t> dataFormat = (uint8_t)extras->data.format;
+      be_val<uint16_t> dataFormat = (uint16_t)extras->data.format;
       decaf_abort(fmt::format("Unexpected voice data format {}", dataFormat));
       //decaf_abort(fmt::format("Unexpected voice data format {}", extras->data.format))
    }
@@ -227,7 +227,7 @@ AXSetVoiceCurrentOffset(AXVoice *voice,
    } else if (extras->data.format == AXVoiceFormat::LPCM8) {
       extras->data.currentOffsetAbs = baseAddress + offset;
    } else {
-      be_val<uint8_t> dataFormat = (uint8_t)extras->data.format;
+      be_val<uint16_t> dataFormat = (uint16_t)extras->data.format;
       decaf_abort(fmt::format("Unexpected voice data format {}", dataFormat));
       //decaf_abort(fmt::format("Unexpected voice data type {}", extras->data.format));
    }
@@ -313,7 +313,7 @@ AXSetVoiceEndOffset(AXVoice *voice,
    } else if (extras->data.format == AXVoiceFormat::LPCM8) {
       extras->data.endOffsetAbs = baseAddress + offset;
    } else {
-      be_val<uint8_t> dataFormat = (uint8_t)extras->data.format;
+      be_val<uint16_t> dataFormat = (uint16_t)extras->data.format;
       decaf_abort(fmt::format("Unexpected voice data format {}", dataFormat));
       //decaf_abort(fmt::format("Unexpected voice data type {}", extras->data.format));
    }
@@ -372,7 +372,7 @@ AXSetVoiceLoopOffset(AXVoice *voice,
    } else if (extras->data.format == AXVoiceFormat::LPCM8) {
       extras->data.loopOffsetAbs = baseAddress + offset;
    } else {
-      be_val<uint8_t> dataFormat = (uint8_t)extras->data.format;
+      be_val<uint16_t> dataFormat = (uint16_t)extras->data.format;
       decaf_abort(fmt::format("Unexpected voice data format {}", dataFormat));
       //decaf_abort(fmt::format("Unexpected voice data type {}", extras->data.format));
    }
@@ -445,7 +445,7 @@ AXSetVoiceOffsets(AXVoice *voice,
       absOffsets.currentOffsetAbs = (samples + offsets->currentOffset) & 0x1fffffff;
       absOffsets.memPageNumber = (samples + offsets->currentOffset) >> 29;
    } else {
-      be_val<uint8_t> offsetDataType = (uint8_t)offsets->dataType;
+      be_val<uint16_t> offsetDataType = (uint16_t)offsets->dataType;
       decaf_abort(fmt::format("Unexpected voice data type {}", offsetDataType));
       //decaf_abort(fmt::format("Unexpected voice data type {}", offsets->dataType));
    }
@@ -607,7 +607,7 @@ setVoiceAddresses(AXVoice *voice,
       extras->adpcm.gain = 0x800;
       voice->syncBits |= internal::AXVoiceSyncBits::Adpcm;
    } else {
-      be_val<uint8_t> offsetFormat = (uint8_t)offsets->format;
+      be_val<uint16_t> offsetFormat = (uint16_t)offsets->format;
       decaf_abort(fmt::format("Unexpected audio data format {}", offsetFormat));
       //decaf_abort(fmt::format("Unexpected audio data format {}", offsets->format));
    }
